requirements.txt
django==4.2.13
djangorestframework==3.15.2
drf_spectacular==0.26.5
django-environ
Pillow==10.1.0
django_cors_headers==4.3.1
requests==2.31.0
django_jalali_date==1.0.2
django-allauth
dj-rest-auth==6.0.0
cryptography>=3.4.7
python_jose==3.3.0
djangorestframework_jwt==1.11.0
djangorestframework_simplejwt==5.3.1
holidays==0.39
kavenegar==1.1.2
django_resized==1.0.2
django_admin_interface==0.28.5
mysqlclient==2.2.4
mysql-connector-python==8.3.0
PyMySQL==1.1.0
qrcode==7.4.2
#djongo==1.3.6
setuptools==70.0.0
gunicorn==20.1.0
dnspython==2.6.1
aiohttp==3.10.3
# django-storages
# boto3
# whitenoise
django_q==1.3.9
pandas==2.2.3
openpyxl==3.1.5
persiantools==5.1.0
django_tinymce==4.1.0
git+https://github.com/shagho/Django-jalali-date-new.git
